[fix] Added error handling for rejected websocket connection
Bug:
The "disconnect" method of the WebsocketConsumer is called after
the "close" method. Termination of a connection which was never
accepted also triggers this method.

Since certain attributes are only set after the connection is accepted,
an error is raised during termination of rejected connection requests

Fix:
Added error handling block to gracefully handle the raised error